requirements.txt
# Only use versions supporting Python 3.5
pillow==7.2.0
python-vlc==3.0.11115
streamdeck==0.7.3
Flask==1.1.2
youtube_dl==2020.11.12